Refactoring: Refined relay package structure
