+++ django_celery_outbox/relay/__init__.py
@@ -1,0 +1,6 @@
+from ._relay import Relay, RelayConfig
+
+__all__ = [
+    'Relay',
+    'RelayConfig',
+]
